Refactor blueprint deeplink handling to use IPC-specific hook
